+++ xpath-to-json-javax/src/main/java/com/github/simy4/xpath/json/spi/JavaxJsonNavigatorSpi.java
@@ -7,11 +7,14 @@ import com.github.simy4.xpath.spi.Effect;
 import com.github.simy4.xpath.spi.NavigatorSpi;
 
 import javax.json.JsonValue;
+import javax.json.spi.JsonProvider;
 
 /**
  * Javax JSON model navigator extension SPI.
  */
 public class JavaxJsonNavigatorSpi implements NavigatorSpi {
+
+    private final JsonProvider jsonProvider = JsonProvider.provider();
 
     @Override
     public boolean canHandle(Object o) {
@@ -25,7 +28,7 @@ public class JavaxJsonNavigatorSpi implements NavigatorSpi {
             throw new IllegalArgumentException("JSON model is not supported");
         }
         final var root = new JavaxJsonRootNode((JsonValue) json);
-        final var navigator = new JavaxJsonNavigator(root);
+        final var navigator = new JavaxJsonNavigator(jsonProvider, root);
         for (var effect : effects) {
             effect.perform(navigator, root);
         }