+++ xpath-to-json-javax/src/test/java/com/github/simy4/xpath/json/navigator/node/JavaxJsonRootNodeTest.java
@@ -6,12 +6,15 @@ import org.junit.jupiter.api.Test;
 import javax.json.JsonArray;
 import javax.json.JsonObject;
 import javax.json.JsonValue;
+import javax.json.spi.JsonProvider;
 import javax.xml.namespace.QName;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class JavaxJsonRootNodeTest {
+
+    private static final JsonProvider jsonProvider = JsonProvider.provider();
 
     private final JsonObject jsonObject = JsonValue.EMPTY_JSON_OBJECT;
     private final JavaxJsonNode rootNode = new JavaxJsonRootNode(jsonObject);
@@ -29,13 +32,13 @@ class JavaxJsonRootNodeTest {
     @Test
     void shouldReplaceRootNodeOnSet() {
         JsonArray array = JsonValue.EMPTY_JSON_ARRAY;
-        rootNode.set(array);
+        rootNode.set(jsonProvider, array);
         assertThat(rootNode.get()).isSameAs(array);
     }
 
     @Test
     void shouldThrowOnRemove() {
-        assertThatThrownBy(rootNode::remove).isInstanceOf(XmlBuilderException.class);
+        assertThatThrownBy(() -> rootNode.remove(jsonProvider)).isInstanceOf(XmlBuilderException.class);
     }
 
 }